Tests fix.

+++ src/MiNET/TestPlugin/NiceLobby/NiceLobbyPlugin.cs
@@ -772,7 +772,7 @@ namespace TestPlugin.NiceLobby
 		[PacketHandler, Send]
 		public Packet AddPlayerHandler(McpeAddPlayer packet, Player player)
 		{
-			if (_playerEntities.Keys.FirstOrDefault(p => p.EntityId == packet.entityIdSelf) != null)
+			if (_playerEntities.Keys.FirstOrDefault(p => (ulong) p.EntityId == packet.entityIdSelf) != null)
 			{
 				return null;
 			}
